requirements.txt
boto3==1.34.0
botocore==1.34.0
Pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-mock==3.12.0
moto==4.2.14
pytest-cov==4.1.0
